+++ lib/distillery/tasks/gen.appup.ex
@@ -25,7 +25,7 @@ defmodule Mix.Tasks.Distillery.Gen.Appup do
   use Mix.Task
 
   alias Distillery.Releases.Shell
-  alias Distillery.Releases.Config
+  alias Distillery.Releases.Conf
   alias Distillery.Releases.Release
   alias Distillery.Releases.Errors
   alias Distillery.Releases.Assembler
@@ -49,7 +49,7 @@ defmodule Mix.Tasks.Distillery.Gen.Appup do
     # load release configuration
     Shell.debug("Loading configuration..")
 
-    case Config.get(opts) do
+    case Conf.get(opts) do
       {:error, {:config, :not_found}} ->
         Shell.error("You are missing a release config file. Run the distillery.init task first")
         System.halt(1)